--- a/splitgraph/objects/dumping.py
+++ b/splitgraph/objects/dumping.py
@@ -1,42 +1,41 @@
-import shlex
-from os import environ
-
-from splitgraph.constants import SPLITGRAPH_META_SCHEMA, PG_HOST, PG_DB, PG_USER, PG_PORT, PG_PWD
+import json
 import logging
-import subprocess
+
+from psycopg2.sql import SQL, Identifier
+from splitgraph.constants import SPLITGRAPH_META_SCHEMA
+from splitgraph.pg_utils import get_full_table_schema, create_table
 
 
 # Utilities to dump objects (SNAP/DIFF) into an external format.
-# This used to be a SQL dump (inspect the schema and write a series of statements that recreate the table)
-# but we now just shell out to pg_dump -- this does require pg_dump locally to be compatible with the
-# driver (pg10).
+# We use a slightly ad hoc format: the schema (JSON) + a null byte + Postgres's copy_to
+# binary format (only contains data). There's probably some scope to make this more optimized, maybe
+# we should look into columnar on-disk formats (Parquet/Avro) but we currently just want to get the objects
+# out of/into postgres as fast as possible.
 
-def dump_object_to_file(object_id, path):
-    # Shell out into pg_dump and use its custom binary dump format to dump the object table.
-    subprocess.check_output(['pg_dump', '-h', PG_HOST, '-d', PG_DB, '-U', PG_USER, '-p', PG_PORT,
-                             '-t', SPLITGRAPH_META_SCHEMA + '.' + object_id,  # Only dump the single table
-                             '-Fc',  # Use the PG compressed binary format
-                             '-f', path], env={**environ, 'PGPASSWORD': PG_PWD})
-    # Note we pass the password to pg_dump with an envvar, which might be considered insecure on systems where one
-    # can inspect other processes' envvars
-    # Though the alternatives are piping it in (which I couldn't get to work since it uses the tty for password inputs
-    # and any information on how to bypass that leads to one of "well you shouldn't do it" and "use pexpect". Or
-    # we could use a .pgpass file but we store pg creds in .sgconfig anyway.
+def dump_object_to_file(conn, object_id, path):
+    with open(path, 'wb') as f:
+        schema = json.dumps(get_full_table_schema(conn, SPLITGRAPH_META_SCHEMA, object_id))
+        f.write(schema.encode('utf-8') + b'\0')
+        with conn.cursor() as cur:
+            cur.copy_expert(SQL("COPY {}.{} TO STDOUT WITH (FORMAT 'binary')")
+                            .format(Identifier(SPLITGRAPH_META_SCHEMA), Identifier(object_id)), f)
 
 
-def load_object_from_file(path):
-    logging.info("Loading objects from %s", path)
-    # Shell out into pg_restore to restore the object table from the archive.
-    # We are basically running some SQL code we downloaded off the internet, so some quick precautions here:
-    # we create only tables that don't exist and only in the splitgraph_meta schema.
-    # We could also run this with -l to list the contents of the archive and raise if something looks dodgy.
-    subprocess.check_output(['pg_restore', path, '-h', PG_HOST, '-d', PG_DB, '-U', PG_USER, '-p', PG_PORT,
-                             '-e',  # exit on error
-                             '-O',  # don't set ownership information
-                             '-x',  # don't set grants
-                             '-n', SPLITGRAPH_META_SCHEMA,  # pg_restore still doesn't support restoring into
-                             # a different schema
-                             # We could pass -t here to only specify a single table but then it wouldn't restore
-                             # its indices/constraints etc.
-                             '--no-data-for-failed-tables',  # don't append into tables that exist
-                             '-Fc'], env={**environ, 'PGPASSWORD': PG_PWD})
+def load_object_from_file(conn, object_id, path):
+    logging.info("Loading %s from %s", object_id, path)
+    with open(path, 'rb') as f:
+        chars = b''
+        # Read until the delimiter separating a JSON schema from the Postgres copy_to dump.
+        # Surely this is buffered?
+        while True:
+            c = f.read(1)
+            if c == b'\0':
+                break
+            chars += c
+
+        schema = json.loads(chars.decode('utf-8'))
+        create_table(conn, SPLITGRAPH_META_SCHEMA, object_id, schema)
+
+        with conn.cursor() as cur:
+            cur.copy_expert(SQL("COPY {}.{} FROM STDIN WITH (FORMAT 'binary')")
+                            .format(Identifier(SPLITGRAPH_META_SCHEMA), Identifier(object_id)), f)