@mildbyte 2018-11-22
Looks like we can't pipe the object directly to S3 (or an S3-like storage) because the API needs to know the size of the object in advance before it's uploaded: https://github.com/minio/mc/issues/2271 https://github.com/minio/minio-py/issues/650