@mildbyte 2018-11-22
rebased on experiment/poetry so that I can play around with alternative S3 clients