@mildbyte 2018-11-23
Piping directly to smart_open -> boto instead of minio: slower, since smart_open's connection initializing takes about 0.18s.

```
Fri Nov 23 11:13:07 2018    push_s3_boto_streaming.cprofile
         36774783 function calls (36168224 primitive calls) in 150.104 seconds
   Ordered by: cumulative time
   List reduced from 1354 to 20 due to restriction <20>
   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
     18/1    0.001    0.000  150.359  150.359 {built-in method builtins.exec}
        1    0.001    0.001  150.359  150.359 <string>:1(<module>)
        1    0.001    0.001  150.358  150.358 /home/mildbyte/splitgraph-opensource/splitgraph/commands/push_pull.py:176(push)
        1    0.001    0.001  147.593  147.593 /home/mildbyte/splitgraph-opensource/splitgraph/objects/loading.py:83(upload_objects)
        1    0.474    0.474  147.587  147.587 /home/mildbyte/splitgraph-opensource/splitgraph/objects/s3.py:22(s3_upload_objects)
      570    0.019    0.000  106.318    0.187 /home/mildbyte/miniconda3/envs/splitgraph-prototype/lib/python3.6/site-packages/smart_open/smart_open_lib.py:114(smart_open)
      570    0.011    0.000  106.177    0.186 /home/mildbyte/miniconda3/envs/splitgraph-prototype/lib/python3.6/site-packages/smart_open/smart_open_lib.py:307(_open_binary_stream)
      570    0.011    0.000  106.138    0.186 /home/mildbyte/miniconda3/envs/splitgraph-prototype/lib/python3.6/site-packages/smart_open/smart_open_lib.py:383(_s3_open_uri)
      570    0.018    0.000  106.123    0.186 /home/mildbyte/miniconda3/envs/splitgraph-prototype/lib/python3.6/site-packages/smart_open/s3.py:60(open)
      570    0.039    0.000  106.099    0.186 /home/mildbyte/miniconda3/envs/splitgraph-prototype/lib/python3.6/site-packages/smart_open/s3.py:349(__init__)
      570    0.038    0.000   55.077    0.097 /home/mildbyte/miniconda3/envs/splitgraph-prototype/lib/python3.6/site-packages/boto3/session.py:265(resource)
     2281    0.060    0.000   43.716    0.019 /home/mildbyte/miniconda3/envs/splitgraph-prototype/lib/python3.6/site-packages/botocore/client.py:312(_api_call)
     2281    0.142    0.000   43.636    0.019 /home/mildbyte/miniconda3/envs/splitgraph-prototype/lib/python3.6/site-packages/botocore/client.py:578(_make_api_call)
      571    0.006    0.000   41.584    0.073 /home/mildbyte/miniconda3/envs/splitgraph-prototype/lib/python3.6/site-packages/boto3/session.py:185(client)
      571    0.032    0.000   41.578    0.073 /home/mildbyte/miniconda3/envs/splitgraph-prototype/lib/python3.6/site-packages/botocore/session.py:762(create_client)
     1710    0.017    0.000   41.440    0.024 /home/mildbyte/miniconda3/envs/splitgraph-prototype/lib/python3.6/site-packages/boto3/resources/factory.py:519(do_action)
     1710    0.066    0.000   41.417    0.024 /home/mildbyte/miniconda3/envs/splitgraph-prototype/lib/python3.6/site-packages/boto3/resources/action.py:62(__call__)
7418/2853    0.079    0.000   40.187    0.014 /home/mildbyte/miniconda3/envs/splitgraph-prototype/lib/python3.6/site-packages/botocore/loaders.py:126(_wrapper)
     2281    0.030    0.000   36.463    0.016 /home/mildbyte/miniconda3/envs/splitgraph-prototype/lib/python3.6/site-packages/botocore/endpoint.py:99(make_request)
     2281    0.034    0.000   36.422    0.016 /home/mildbyte/miniconda3/envs/splitgraph-prototype/lib/python3.6/site-packages/botocore/endpoint.py:130(_send_request)
```