@mildbyte 2018-11-23
Minio, multithreaded (4 threads):

570 small objects (a lot of waiting on the pg connection since one is shared between all threads): 14.8s upload (~2x faster than wo multithreading)

```
Fri Nov 23 12:29:43 2018    push_minio_streaming_4threads.cprofile
         394700 function calls (394474 primitive calls) in 14.880 seconds
   Ordered by: cumulative time
   List reduced from 579 to 20 due to restriction <20>
   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
        1    0.000    0.000   14.883   14.883 {built-in method builtins.exec}
        1    0.001    0.001   14.883   14.883 <string>:1(<module>)
        1    0.001    0.001   14.882   14.882 /home/mildbyte/splitgraph-opensource/splitgraph/commands/push_pull.py:176(push)
        1    0.001    0.001   12.450   12.450 /home/mildbyte/splitgraph-opensource/splitgraph/objects/loading.py:83(upload_objects)
        1    0.000    0.000   12.437   12.437 /home/mildbyte/splitgraph-opensource/splitgraph/objects/s3.py:23(s3_upload_objects)
      605   12.382    0.020   12.382    0.020 {method 'acquire' of '_thread.lock' objects}
        1    0.000    0.000   12.381   12.381 /home/mildbyte/miniconda3/envs/splitgraph-prototype/lib/python3.6/concurrent/futures/_base.py:610(__exit__)
        1    0.000    0.000   12.381   12.381 /home/mildbyte/miniconda3/envs/splitgraph-prototype/lib/python3.6/concurrent/futures/thread.py:146(shutdown)
        4    0.000    0.000   12.381    3.095 /home/mildbyte/miniconda3/envs/splitgraph-prototype/lib/python3.6/threading.py:1024(join)
        4    0.000    0.000   12.381    3.095 /home/mildbyte/miniconda3/envs/splitgraph-prototype/lib/python3.6/threading.py:1062(_wait_for_tstate_lock)
        1    0.024    0.024    1.964    1.964 /home/mildbyte/splitgraph-opensource/splitgraph/commands/push_pull.py:18(_get_required_snaps_objects)
     1179    1.747    0.001    1.870    0.002 {method 'execute' of 'psycopg2.extensions.cursor' objects}
      571    0.017    0.000    0.989    0.002 /home/mildbyte/splitgraph-opensource/splitgraph/meta_handler/images.py:76(add_new_image)
        4    0.001    0.000    0.425    0.106 /home/mildbyte/miniconda3/envs/splitgraph-prototype/lib/python3.6/site-packages/psycopg2/extras.py:1181(execute_batch)
      576    0.023    0.000    0.213    0.000 /home/mildbyte/splitgraph-opensource/splitgraph/meta_handler/common.py:112(insert)
        2    0.000    0.000    0.189    0.094 /home/mildbyte/splitgraph-opensource/splitgraph/meta_handler/objects.py:54(register_object_locations)
     3434    0.064    0.000    0.186    0.000 /home/mildbyte/miniconda3/envs/splitgraph-prototype/lib/python3.6/site-packages/psycopg2/sql.py:119(as_string)
       24    0.003    0.000    0.182    0.008 /home/mildbyte/miniconda3/envs/splitgraph-prototype/lib/python3.6/site-packages/psycopg2/extras.py:1202(<listcomp>)
     2280    0.053    0.000    0.179    0.000 {method 'mogrify' of 'psycopg2.extensions.cursor' objects}
        1    0.000    0.000    0.146    0.146 /home/mildbyte/splitgraph-opensource/splitgraph/meta_handler/objects.py:47(register_tables)
```

1M SNAP, 1K DIFF (no gzipping): 13.2s upload, 20s download

100K SNAP, 1K DIFF (no gzipping): 1.8s upload, 2.3s download